+++ src/cursor-doc/clojure-lexer.ts
@@ -168,7 +168,7 @@ export interface ScannerState {
 export class Scanner {
   state: ScannerState = { inString: false };
 
-  constructor(private maxLength: number) {}
+  constructor(public maxLength: number) {}
 
   processLine(line: string, state: ScannerState = this.state) {
     const tks: Token[] = [];