+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.373] - 2023-06-27
+
+- [Re-indent non-complete code, when pasting or formatting selection](https://github.com/BetterThanTomorrow/calva/issues/1709)
+- Fix: [Selection, closing brackets, fails in mysterious ways](https://github.com/BetterThanTomorrow/calva/issues/2232)
+
 ## [2.0.372] - 2023-06-23
 
 - Fix: [Trailing space is removed unconditionally from pasted text](https://github.com/BetterThanTomorrow/calva/issues/2229)